Refactored graphql section of the code.
