+++ Gopkg.lock
@@ -2,10 +2,16 @@
 
 
 [[projects]]
-  name = "github.com/bradleyjkemp/memmap"
+  name = "cloud.google.com/go"
+  packages = ["compute/metadata"]
+  revision = "767c40d6a2e058483c25fa193e963a22da17236d"
+  version = "v0.18.0"
+
+[[projects]]
+  name = "github.com/dgrijalva/jwt-go"
   packages = ["."]
-  revision = "8469122a29ebfa6dc14416f742870ebf50370cc9"
-  version = "v0.2.0"
+  revision = "dbeaa9332f19a944acb5736b4456cfcc02140e29"
+  version = "v3.1.0"
 
 [[projects]]
   branch = "master"
@@ -24,6 +30,24 @@
   name = "github.com/google/go-querystring"
   packages = ["query"]
   revision = "53e6ce116135b80d037921a7fdd5138cf32d7a8a"
+
+[[projects]]
+  name = "github.com/gorilla/context"
+  packages = ["."]
+  revision = "1ea25387ff6f684839d82767c1733ff4d4d15d0a"
+  version = "v1.1"
+
+[[projects]]
+  name = "github.com/gorilla/securecookie"
+  packages = ["."]
+  revision = "667fe4e3466a040b780561fe9b51a83a3753eefc"
+  version = "v1.1"
+
+[[projects]]
+  name = "github.com/gorilla/sessions"
+  packages = ["."]
+  revision = "ca9ada44574153444b00d3fd9c8559e4cc95f896"
+  version = "v1.1"
 
 [[projects]]
   name = "github.com/graphql-go/graphql"
@@ -51,6 +75,114 @@
 
 [[projects]]
   branch = "master"
+  name = "github.com/jinzhu/copier"
+  packages = ["."]
+  revision = "db4671f3a9b8df855e993f7c94ec5ef1ffb0a23b"
+
+[[projects]]
+  name = "github.com/jinzhu/gorm"
+  packages = ["."]
+  revision = "5174cc5c242a728b435ea2be8a2f7f998e15429b"
+  version = "v1.0"
+
+[[projects]]
+  branch = "master"
+  name = "github.com/jinzhu/inflection"
+  packages = ["."]
+  revision = "1c35d901db3da928c72a72d8458480cc9ade058f"
+
+[[projects]]
+  branch = "master"
+  name = "github.com/mrjones/oauth"
+  packages = ["."]
+  revision = "3f67d9c274355678b2f9844b08d643e2f9213340"
+
+[[projects]]
+  branch = "master"
+  name = "github.com/qor/assetfs"
+  packages = ["."]
+  revision = "ff57fdc13a148d39c3e930cb27c7eaf82e376891"
+
+[[projects]]
+  branch = "master"
+  name = "github.com/qor/auth"
+  packages = [
+    ".",
+    "auth_identity",
+    "claims",
+    "providers/facebook",
+    "providers/github",
+    "providers/google",
+    "providers/password",
+    "providers/password/encryptor",
+    "providers/password/encryptor/bcrypt_encryptor",
+    "providers/twitter"
+  ]
+  revision = "d0a475c04fd9ead5301ed7de9cc65776f58a4c87"
+
+[[projects]]
+  branch = "master"
+  name = "github.com/qor/mailer"
+  packages = [
+    ".",
+    "logger"
+  ]
+  revision = "1e6ac71069556a146390b35dea415dc1c383f430"
+
+[[projects]]
+  branch = "master"
+  name = "github.com/qor/middlewares"
+  packages = ["."]
+  revision = "781378b69454a545d08058d8154aca40d079e7ab"
+
+[[projects]]
+  name = "github.com/qor/qor"
+  packages = [
+    ".",
+    "utils"
+  ]
+  revision = "d1402d1caffb8f9c05d73fa5bfaae2853c4191fb"
+  version = "v1.0"
+
+[[projects]]
+  branch = "master"
+  name = "github.com/qor/redirect_back"
+  packages = ["."]
+  revision = "b4161ed6f84819fb3a2d13d5c9e6872ef76d8128"
+
+[[projects]]
+  name = "github.com/qor/render"
+  packages = ["."]
+  revision = "ab08a9cb69bef67462dcfea4a721d2798f53609f"
+  version = "v1.0"
+
+[[projects]]
+  name = "github.com/qor/responder"
+  packages = ["."]
+  revision = "ecae0be66c1a6a441001b0b5eeef51f8bc27d42a"
+  version = "v1.0"
+
+[[projects]]
+  branch = "master"
+  name = "github.com/qor/session"
+  packages = [
+    ".",
+    "gorilla",
+    "manager"
+  ]
+  revision = "8206b0adab706a6ef3ee6fabba2584d34429a26a"
+
+[[projects]]
+  branch = "master"
+  name = "golang.org/x/crypto"
+  packages = [
+    "bcrypt",
+    "blowfish"
+  ]
+  revision = "1875d0a70c90e57f11972aefd42276df65e895b9"
+
+[[projects]]
+  branch = "master"
   name = "golang.org/x/net"
   packages = [
     "context",
@@ -63,23 +195,42 @@
   name = "golang.org/x/oauth2"
   packages = [
     ".",
-    "internal"
+    "facebook",
+    "google",
+    "internal",
+    "jws",
+    "jwt"
   ]
   revision = "30785a2c434e431ef7c507b54617d6a951d5f2b4"
 
 [[projects]]
   name = "google.golang.org/appengine"
   packages = [
+    ".",
     "internal",
+    "internal/app_identity",
     "internal/base",
     "internal/datastore",
     "internal/log",
+    "internal/modules",
     "internal/remote_api",
     "internal/urlfetch",
     "urlfetch"
   ]
   revision = "150dc57a1b433e64154302bdc40b6bb8aefa313a"
   version = "v1.0.0"
+
+[[projects]]
+  branch = "v3"
+  name = "gopkg.in/alexcesaro/quotedprintable.v3"
+  packages = ["."]
+  revision = "2caba252f4dc53eaf6b553000885530023f54623"
+
+[[projects]]
+  name = "gopkg.in/gomail.v2"
+  packages = ["."]
+  revision = "41f3572897373c5538c50a2402db15db079fa4fd"
+  version = "2.0.0"
 
 [[projects]]
   branch = "v2"
@@ -96,6 +247,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "4b8cae75a9666a2db4d167070432d5a1ba907c81e116c16c9a210c2f15bfc3d0"
+  inputs-digest = "86ba5642087cb16ac2994ea8bcc283b510525690aa4673d936238430cf721a40"
   solver-name = "gps-cdcl"
   solver-version = 1
